+++ packages/camera/example/lib/main.dart
@@ -200,7 +200,7 @@ class _CameraExampleHomeState extends State<CameraExampleHome> {
     if (controller != null) {
       await controller.dispose();
     }
-    controller = CameraController(cameraDescription, ResolutionPreset.high);
+    controller = CameraController(cameraDescription, ResolutionPreset.medium);
 
     // If the controller is updated then update the UI.
     controller.addListener(() {