+++ packages/camera/lib/camera.dart
@@ -16,19 +16,31 @@ final MethodChannel _channel = const MethodChannel('plugins.flutter.io/camera');
 
 enum CameraLensDirection { front, back, external }
 
-enum ResolutionPreset { low, medium, high }
+/// Affect the quality of video recording and image capture:
+/// - veryLow = 352x288 (CIF)
+/// - low = 640x480 on iOS, 720x480 on Android (480p)
+/// - medium = 1280x720 (720p)
+/// - high = 1920x1080 (1080p/HD)
+/// - veryHigh = 3840x2160 (2160p/4K)
+///
+/// If a preset is not available on the camera being used a preset of lower quality will be selected automatically.
+enum ResolutionPreset { veryLow, low, medium, high, veryHigh }
 
 typedef onLatestImageAvailable = Function(CameraImage image);
 
 /// Returns the resolution preset as a String.
 String serializeResolutionPreset(ResolutionPreset resolutionPreset) {
   switch (resolutionPreset) {
+    case ResolutionPreset.veryHigh:
+      return 'veryHigh';
     case ResolutionPreset.high:
       return 'high';
     case ResolutionPreset.medium:
       return 'medium';
     case ResolutionPreset.low:
       return 'low';
+    case ResolutionPreset.veryLow:
+      return 'veryLow';
   }
   throw ArgumentError('Unknown ResolutionPreset value');
 }